bump Go to 1.21, compress, push tag attachments
This only builds for PRs and drops the artifact upload, and publishes
attachements for tags instead, so that each individual file can be
downloaded from a release page on GitHub.
We should figure out if we can share the u-root commit to use between
the two actions, or find some other solution.

Signed-off-by: Daniel Maslowski <info@orangecms.org>